Added BoostLibraryIncludes.cmake module for generating library includes.
This module sets BOOST_LIBRARY_INCLUDES variable to the list of all library
includes. The module detects the monolithic include directory, which is present
in the official monolithic Boost distribution and after `b2 headers`, and
returns it instead when present.

This module is useful when implementing CMake configuration checks that rely
on Boost libraries. Configuration checks cannot use CMake targets defined
by libraries because at the point of configuration these targets are not yet
defined.