Avoid installing libraries multiple times or issueing wrong message
Some libraries call `boost_install` themselves which adds
`INSTALL_INTERFACE` to their `INTERFACE_INCLUDE_DIRETORIES` property.
That then makes the install check in `__boost_auto_install` fail and
print a message that the library won't be installed while it will be.

Use a label to detect targets for which `boost_install` was already
called and skip the logic in `__boost_auto_install` for those.